+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Issue [#11](https://github.com/m-marini/wheellycam/issues/11): Add GUI
+- Issue [#13](https://github.com/m-marini/wheellycam/issues/13): Add pause button
+
 ## [0.2.0] 2024-09-02
 
 ### Added